Adjust size of op code menu and add ellipsis
