+++ src/components/performance/PerfCharts.tsx
@@ -1,4 +1,4 @@
-import { FC, useEffect, useMemo, useState } from 'react';
+import { FC } from 'react';
 import PerfDeviceKernelDurationChart from './PerfDeviceKernelDurationChart';
 import PerfDeviceKernelRuntimeChart from './PerfDeviceKernelRuntimeChart';
 import PerfCoreCountUtilizationChart from './PerfCoreCountUtilizationChart';
@@ -6,91 +6,50 @@ import PerfOperationKernelUtilizationChart from './PerfOperationKernelUtilizatio
 import PerfKernelDurationUtilizationChart from './PerfKernelDurationUtilizationChart';
 import PerfOperationTypesChart from './PerfOperationTypesChart';
 import PerfOpCountVsRuntimeChart from './PerfOpCountVsRuntimeChart';
-import PerfChartFilter from './PerfChartFilter';
-import { MARKER_COLOURS, Marker, PerfTableRow } from '../../definitions/PerfTable';
+import { Marker, PerfTableRow } from '../../definitions/PerfTable';
 import 'styles/components/PerfCharts.scss';
 
 interface PerfChartsProps {
     perfData: PerfTableRow[];
     maxCores: number;
+    opCodeOptions: Marker[];
+    selectedOpCodes: Marker[];
 }
 
-const PerfCharts: FC<PerfChartsProps> = ({ perfData, maxCores }) => {
-    const opCodeOptions = useMemo(
-        () =>
-            [
-                ...new Set(
-                    perfData?.map((row) => row.raw_op_code).filter((opCode): opCode is string => opCode !== undefined),
-                ).values(),
-            ]
-                .sort()
-                .map((opCode, index) => ({
-                    opCode,
-                    colour: MARKER_COLOURS[index],
-                })),
-        [perfData],
-    );
-
-    const [filteredPerfData, setFilteredPerfData] = useState<PerfTableRow[]>([]);
-    const [selectedOpCodes, setSelectedOpCodes] = useState<Marker[]>(opCodeOptions);
-
-    useEffect(() => {
-        setFilteredPerfData(
-            perfData
-                ?.filter((row) =>
-                    selectedOpCodes.length
-                        ? selectedOpCodes.map((selected) => selected.opCode).includes(row.raw_op_code ?? '')
-                        : false,
-                )
-                .sort((a, b) => (a.raw_op_code ?? '').localeCompare(b.raw_op_code ?? '')) || [],
-        );
-    }, [selectedOpCodes, perfData]);
-
-    useEffect(() => {
-        setSelectedOpCodes(opCodeOptions);
-    }, [opCodeOptions]);
-
+const PerfCharts: FC<PerfChartsProps> = ({ perfData, maxCores, opCodeOptions, selectedOpCodes }) => {
     return (
-        <div className='charts-container'>
-            <PerfChartFilter
-                opCodeOptions={opCodeOptions}
+        <div className='charts'>
+            <PerfOpCountVsRuntimeChart
+                data={perfData}
                 selectedOpCodes={selectedOpCodes}
-                updateOpCodes={setSelectedOpCodes}
             />
 
-            <div className='charts'>
-                <PerfOpCountVsRuntimeChart
-                    data={filteredPerfData}
-                    selectedOpCodes={selectedOpCodes}
-                />
+            <PerfDeviceKernelRuntimeChart
+                data={perfData}
+                maxCores={maxCores}
+            />
 
-                <PerfDeviceKernelRuntimeChart
-                    data={filteredPerfData}
-                    maxCores={maxCores}
-                />
+            <PerfDeviceKernelDurationChart data={perfData} />
 
-                <PerfDeviceKernelDurationChart data={filteredPerfData} />
+            <PerfCoreCountUtilizationChart
+                data={perfData}
+                maxCores={maxCores}
+            />
 
-                <PerfCoreCountUtilizationChart
-                    data={filteredPerfData}
-                    maxCores={maxCores}
-                />
+            <PerfOperationKernelUtilizationChart
+                data={perfData}
+                maxCores={maxCores}
+            />
 
-                <PerfOperationKernelUtilizationChart
-                    data={filteredPerfData}
-                    maxCores={maxCores}
-                />
+            <PerfKernelDurationUtilizationChart
+                data={perfData}
+                maxCores={maxCores}
+            />
 
-                <PerfKernelDurationUtilizationChart
-                    data={filteredPerfData}
-                    maxCores={maxCores}
-                />
-
-                <PerfOperationTypesChart
-                    data={filteredPerfData}
-                    opCodes={opCodeOptions}
-                />
-            </div>
+            <PerfOperationTypesChart
+                data={perfData}
+                opCodes={opCodeOptions}
+            />
         </div>
     );
 };
