+++ src/scss/components/PerfCharts.scss
@@ -7,4 +7,8 @@
     grid-template-columns: 1fr minmax(min-content, 750px) 1fr;
     position: relative;
     gap: 60px;
+
+    &.has-comparison {
+        grid-template-columns: 1fr minmax(min-content, 750px) minmax(min-content, 750px);
+    }
 }
